--- CardPayments/src/main/java/com/paypal/android/card/api/ConfirmPaymentSourceResponse.kt
@@ -1,13 +1,0 @@
-package com.paypal.android.card.api
-
-import com.paypal.android.card.model.PaymentSource
-import com.paypal.android.card.model.PurchaseUnit
-import com.paypal.android.corepayments.OrderStatus
-
-internal data class ConfirmPaymentSourceResponse(
-    val orderID: String,
-    val status: OrderStatus? = null,
-    val payerActionHref: String? = null,
-    val paymentSource: PaymentSource? = null,
-    val purchaseUnits: List<PurchaseUnit>? = null
-)
