--- CardPayments/src/main/java/com/paypal/android/card/api/CreateOrderResponse.kt
@@ -1,5 +1,0 @@
-package com.paypal.android.card.api
-
-import com.paypal.android.corepayments.OrderStatus
-
-internal data class CreateOrderResponse(val orderID: String, val status: OrderStatus)
