--- CardPayments/src/main/java/com/paypal/android/card/api/GetOrderRequest.kt
@@ -1,3 +1,0 @@
-package com.paypal.android.card.api
-
-internal data class GetOrderRequest(val orderId: String)
