--- CardPayments/src/main/java/com/paypal/android/card/model/Amount.kt
@@ -1,13 +1,0 @@
-package com.paypal.android.card.model
-
-import com.paypal.android.corepayments.PaymentsJSON
-
-data class Amount(
-    val currencyCode: String?,
-    val value: String?
-) {
-    internal constructor(json: PaymentsJSON) : this(
-        json.optString("currency_code"),
-        json.optString("value")
-    )
-}
