--- CardPayments/src/main/java/com/paypal/android/card/model/Payee.kt
@@ -1,7 +1,0 @@
-package com.paypal.android.card.model
-
-import com.paypal.android.corepayments.PaymentsJSON
-
-data class Payee(val emailAddress: String) {
-    internal constructor(json: PaymentsJSON) : this(json.optString("email_address") ?: "")
-}
