--- CardPayments/src/main/java/com/paypal/android/card/threedsecure/SCA.kt
@@ -1,5 +1,0 @@
-package com.paypal.android.card.threedsecure
-
-enum class SCA {
-    SCA_ALWAYS, SCA_WHEN_REQUIRED
-}
