+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardLifeCycleObserver.kt
@@ -1,0 +1,21 @@
+package com.paypal.android.cardpayments
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.lifecycle.DefaultLifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+
+internal class CardLifeCycleObserver(private val cardClient: CardClient) : DefaultLifecycleObserver {
+
+    override fun onResume(owner: LifecycleOwner) {
+        super.onResume(owner)
+
+        val activity = when (owner) {
+            is FragmentActivity -> owner
+            is Fragment -> owner.requireActivity()
+            else -> null
+        }
+
+        activity?.let { cardClient.handleBrowserSwitchResult(it) }
+    }
+}
