+++ CardPayments/src/main/java/com/paypal/android/cardpayments/api/CreateOrderResponse.kt
@@ -1,0 +1,5 @@
+package com.paypal.android.cardpayments.api
+
+import com.paypal.android.corepayments.OrderStatus
+
+internal data class CreateOrderResponse(val orderID: String, val status: OrderStatus)
