+++ CardPayments/src/main/java/com/paypal/android/cardpayments/model/Amount.kt
@@ -1,0 +1,13 @@
+package com.paypal.android.cardpayments.model
+
+import com.paypal.android.corepayments.PaymentsJSON
+
+data class Amount(
+    val currencyCode: String?,
+    val value: String?
+) {
+    internal constructor(json: PaymentsJSON) : this(
+        json.optString("currency_code"),
+        json.optString("value")
+    )
+}
