+++ CardPayments/src/main/java/com/paypal/android/cardpayments/threedsecure/SCA.kt
@@ -1,0 +1,5 @@
+package com.paypal.android.cardpayments.threedsecure
+
+enum class SCA {
+    SCA_ALWAYS, SCA_WHEN_REQUIRED
+}
