+++ Demo/src/main/java/com/paypal/android/data/card/TestCard.kt
@@ -1,5 +1,5 @@
 package com.paypal.android.data.card
 
-import com.paypal.android.card.Card
+import com.paypal.android.cardpayments.Card
 
 data class TestCard(val name: String, val card: Card)
