+++ Demo/src/main/java/com/paypal/android/ui/card/CardFragment.kt
@@ -13,12 +13,12 @@ import com.paypal.android.R
 import com.paypal.android.api.model.CreateOrderRequest
 import com.paypal.android.api.model.Payee
 import com.paypal.android.api.services.SDKSampleServerApi
-import com.paypal.android.card.ApproveOrderListener
-import com.paypal.android.card.Card
-import com.paypal.android.card.CardClient
-import com.paypal.android.card.CardRequest
-import com.paypal.android.card.model.CardResult
-import com.paypal.android.card.threedsecure.SCA
+import com.paypal.android.cardpayments.ApproveOrderListener
+import com.paypal.android.cardpayments.Card
+import com.paypal.android.cardpayments.CardClient
+import com.paypal.android.cardpayments.CardRequest
+import com.paypal.android.cardpayments.model.CardResult
+import com.paypal.android.cardpayments.threedsecure.SCA
 import com.paypal.android.corepayments.Address
 import com.paypal.android.corepayments.CoreConfig
 import com.paypal.android.corepayments.PayPalSDKError
